lucet-validate: print out type sig of all imports, start func sig
